Reorder #if 0/#endif block to avoid splitting a basic block across the #if 0.
The bracing convention caused a #if 0/#endif to exclude the close of one
block and the open of another.  This compiled correctly, but confused
other tools that expect to see a block fully present or fully absent.
Move the ending brace of the preceding block out of the #if 0 and the
ending brace of the excluded block into the #if 0 to fix that.