Server: always drop privileges, even in foreground mode.
Foreground mode is often used as a debugging aid for live
configurations, so the server may be started with root privileges to let
it bind to its native port.  Since the server already has the ability to
drop root privileges, use that ability to make foreground mode a little
safer.