Cleanup: cast socket to long before logging it.
The SilcSocket type varies in size, depending on the host platform.  On
Unix systems, it is an int, which is 32-bits on all interesting
platforms.  On Windows, it is a SOCKET, which is also 32-bits.  However,
on Symbian, SilcSocket is a void pointer.  Changing %lu to %u may not
work in all cases due to this size difference, so instead cast the
socket to a long unsigned, which should avoid loss of precision on all
platforms and still satisfy the %lu format specifier.