Don't count detached users in server local user count.
This commit decrements the count of users when a detach occurs, and
increments the count when a user re-attaches.

This is done as current silc routers have a bug where only a detached
users previous home server is not properly pro-actively notified of a
user resuming elsewhere unless the user resumes on a router server.

The issue with non-router-server resume notification will be fixed
seperately.  This change allows member servers to operate reliably in
the event that there is a router with the aforementioned bug in their
silcnet.