silcd: Fixed SIGUSR1 signal handling
Fixed the SIGUSR1 signal handling which can be used to dump the server
internals to /tmp.

Changed also End of Stream handling to handle NULL idata pointer instead
of ignoring the EOS in case it is NULL.

Changed also the DETACH timeout handling to use the packet stream
directly instead of looking up client in the callback.