Enable higher security messages MACs
The code to include the source and destination ID in the message payload
MAC has been there for a long time but the use of it has been disabled.
This commit enables it but preserves the backwards support for those
clients unable to verify the MACs.  The support for the newer MACs
have been there for several years.