silccore: packet injection and stream wrap improvements
Add silc_packet_stream_inject to allow injecting of packets to the
specified packet stream.

Add support for specifying the source and destination ids for the
wrapped packet stream allowing to use them in packet sending and using
them in packet reception to take only packets with the specified ids.

The semantics of CAN_WRITE and CAN_READ of wrapped packet stream coder
function has been changed to allow the coder to filter out packets it
does not want or to handle errors in coding.