silcclient: check packet type as responder before starting AKE
Do not immediately start the private message key autonegotiation as
responder when a packet comes in but wait until it is decoded from the
private message payload so that responder can properly set up the SKE
properties and start the SKE in proper state.  Initiator is allowed to
start SKE with SILC_PACKET_KEY_EXCHANGE at any time, including when a key
already exists and it would be error to expect that initiator should have
sent SILC_PACKET_KEY_EXCHANGE_1 just because key exists in responder side.